revert: shifts completetly back to monolithic
