+++ vercel.json
@@ -19,7 +19,7 @@
         {
             "src": "/app/api.py",
             "use": "@vercel/python",
-            "config": {"runtime": "python 3.11"}
+            "config": {"runtime": "python 3.12"}
         }
     ],
     "routes": [
